// src/modules/personalize/models/requests/CustomerVisitorRequest.ts
export interface CustomerVisitorRequest {
  day02?: number;
  day01?: number;
  day03?: number;
  day04?: number;
  day05?: number;
  day06?: number;
  day07?: number;
  day08?: number;
  day09?: number;
  day10?: number;
  day11?: number;
  day12?: number;
  day13?: number;
  day14?: number;
  day15?: number;
  day16?: number;
  day17?: number;
  day18?: number;
  day19?: number;
  day20?: number;
  day21?: number;
  day22?: number;
  day23?: number;
  day24?: number;
  day25?: number;
  day26?: number;
  day27?: number;
  day28?: number;
  day29?: number;
  day30?: number;
  day31?: number;
  month: number;
  year: number;
  storeId: number;
  assigneeCode: string;
  assigneeName: string;
  source: string;
}