// src/modules/personalize/models/responses/CustomerVisitorResponse.ts
export interface CustomerVisitorResponse {
  day01: number | undefined;
  day02: number | undefined;
  day03: number | undefined;
  day04: number | undefined;
  day05: number | undefined;
  day06: number | undefined;
  day07: number | undefined;
  day08: number | undefined;
  day09: number | undefined;
  day10: number | undefined;
  day11: number | undefined;
  day12: number | undefined;
  day13: number | undefined;
  day14: number | undefined;
  day15: number | undefined;
  day16: number | undefined;
  day17: number | undefined;
  day18: number | undefined;
  day19: number | undefined;
  day20: number | undefined;
  day21: number | undefined;
  day22: number | undefined;
  day23: number | undefined;
  day24: number | undefined;
  day25: number | undefined;
  day26: number | undefined;
  day27: number | undefined;
  day28: number | undefined;
  day29: number | undefined;
  day30: number | undefined;
  day31: number | undefined;
  assigneeCode: string;
  assigneeName: string;
  createdAt: string;
  createdBy: string;
  month: number;
  storeId: number;
  updatedAt: string;
  updatedBy: string;
  year: number;
}